Added Upcoming View functionallity and changed certain UI Elements to look better.
