+++ src/main/java/seedu/address/ui/panel/calendar/UpcomingView.java
@@ -1,19 +1,18 @@
 package seedu.address.ui.panel.calendar;
 
-import javafx.concurrent.Task;
-import javafx.concurrent.WorkerStateEvent;
-import javafx.event.EventHandler;
+import java.util.List;
+
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
+import javafx.scene.layout.GridPane;
 import javafx.scene.layout.Region;
-import javafx.scene.layout.StackPane;
 import javafx.scene.layout.VBox;
+
 import seedu.address.model.CalendarDate;
 import seedu.address.model.events.EventSource;
 import seedu.address.ui.UiPart;
 import seedu.address.ui.card.UpcomingEventCard;
 
-import java.util.List;
 
 /**
  * An UI that represents the upcoming event or task displayed to the user.
@@ -24,10 +23,10 @@ public class UpcomingView extends UiPart<Region> {
     private Integer totalDisplays;
 
     @FXML
-    private Label upcomingViewTitle;
+    private GridPane upcomingViewBase;
 
     @FXML
-    private StackPane upcomingPanel;
+    private Label upcomingViewTitle;
 
     @FXML
     private VBox upcomingList;
@@ -36,7 +35,6 @@ public class UpcomingView extends UiPart<Region> {
         super(FXML);
         CalendarDate calendarDate = CalendarDate.now();
         upcomingViewTitle.setText("Upcoming in " + calendarDate.getEnglishMonth());
-        changeTotalDisplays();
     }
 
     /**
@@ -46,8 +44,8 @@ public class UpcomingView extends UiPart<Region> {
     public void eventChange(List<EventSource> events) {
         int index = 0;
         upcomingList.getChildren().clear();
-        for(EventSource event : events) {
-            if(index >= totalDisplays) {
+        for (EventSource event : events) {
+            if (index >= totalDisplays) {
                 break;
             }
             UpcomingEventCard eventCard = new UpcomingEventCard(event);
@@ -56,29 +54,13 @@ public class UpcomingView extends UiPart<Region> {
         }
     }
 
-    private void changeTotalDisplays() {
-        Task<Void> sleeper = new Task<Void>() {
-            @Override
-            protected Void call() throws Exception {
-                try {
-                    Thread.sleep(10);
-                } catch (InterruptedException e) {
-                    throw new Exception(e.getMessage());
-                }
-                return null;
-            }
-        };
-        sleeper.setOnSucceeded(new EventHandler<WorkerStateEvent>() {
-            @Override
-            public void handle(WorkerStateEvent event) {
-                totalDisplays = (int) upcomingPanel.getHeight() / 39 - 1;
-            }
-        });
-        new Thread(sleeper).start();
-    }
-
+    /**
+     * Re-sizes the upcoming to fit the amount of displays of the UpcomingEventCard.
+     *
+     * @param events The list of events to change the upcoming view.
+     */
     public void resizeUpcomingView(List<EventSource> events) {
-        changeTotalDisplays();
+        totalDisplays = (int) (upcomingViewBase.getHeight() / UpcomingEventCard.UPCOMING_CARD_HEIGHT) - 2;
         eventChange(events);
     }
 }
